\name{accrualReport}
\alias{accrualReport}
\title{Accrual Report}
\usage{
  accrualReport(Minor, Major = rep("", length(Minor)),
    MajorLabel = "", MinorLabel = "", EntryDate1 = NULL,
    EntryDate2 = NULL, EntryDateLabel = "", EntryDate1cap,
    EntryDate2cap, dateRange, dateformat = "y-m-d",
    targetN, targetDate = NULL, panel = "randomized",
    hdotchart = 6, append = TRUE)
}
\arguments{
  \item{Minor}{something minor NEEDDOC}

  \item{Major}{NEEDDOC}

  \item{MajorLabel}{NEEDDOC}

  \item{MinorLabel}{NEEDDOC}

  \item{EntryDate1}{NEEDDOC}

  \item{EntryDate2}{NEEDDOC}

  \item{EntryDateLabel}{NEEDDOC}

  \item{EntryDate1cap}{NEEDDOC}

  \item{EntryDate2cap}{NEEDDOC}

  \item{dateRange}{NEEDDOC}

  \item{dateformat}{NEEDDOC}

  \item{targetN}{NEEDDOC}

  \item{targetDate}{NEEDDOC}

  \item{panel}{NEEDDOC}

  \item{hdotchart}{NEEDDOC}

  \item{append}{NEEDDOC}
}
\value{
  return something
}
\description{
  quick summary
}
\details{
  detailed description
}
\examples{
1
}

