\name{aeReport}
\alias{aeReport}
\title{AE Report}
\usage{
  aeReport(data = NULL, vars, treat, time,
    times = sort(unique(time)), id = NULL,
    plotprop = FALSE, plotkm = TRUE, etimedata = NULL,
    tables = TRUE, times.tables = times,
    forceBinary = FALSE, ylim = c(0, 1), h = 5, w = 5,
    digits = 3)
}
\arguments{
  \item{data}{NEEDDOC}

  \item{vars}{NEEDDOC}

  \item{treat}{NEEDDOC}

  \item{time}{NEEDDOC}

  \item{times}{NEEDDOC}

  \item{id}{NEEDDOC}

  \item{plotprop}{NEEDDOC}

  \item{plotkm}{NEEDDOC}

  \item{etimedata}{NEEDDOC}

  \item{tables}{NEEDDOC}

  \item{times.tables}{NEEDDOC}

  \item{forceBinary}{NEEDDOC}

  \item{ylim}{NEEDDOC}

  \item{h}{NEEDDOC}

  \item{w}{NEEDDOC}

  \item{digits}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

