\name{displayFreq}
\alias{displayFreq}
\title{Display AE Frequencies}
\usage{
  displayFreq(dataframe, subjectVar, minorVar, majorVar,
    occurrenceVar, stratVar, denomSub = NULL,
    fileName = NULL, labelLen = 10, pvalue = 0.05,
    keepPvalue = 0.5, minDisplayNum = 2, majorGrid = NULL,
    minorGrid = NULL, plotGrid = TRUE, gridDig = 0,
    titleOffsetKoef = 10, titleOffsetX = 0,
    minorToMajorKoef = 5, sparseKoef = 1.5, graphWidth = 8,
    graphHeight = 11, gridCex = 0.5, labelCex = 0.5,
    titles = NULL)
}
\arguments{
  \item{dataframe}{NEEDDOC}

  \item{subjectVar}{NEEDDOC}

  \item{minorVar}{NEEDDOC}

  \item{majorVar}{NEEDDOC}

  \item{occurrenceVar}{NEEDDOC}

  \item{stratVar}{NEEDDOC}

  \item{denomSub}{NEEDDOC}

  \item{fileName}{NEEDDOC}

  \item{labelLen}{NEEDDOC}

  \item{pvalue}{NEEDDOC}

  \item{keepPvalue}{NEEDDOC}

  \item{minDisplayNum}{NEEDDOC}

  \item{majorGrid}{NEEDDOC}

  \item{minorGrid}{NEEDDOC}

  \item{plotGrid}{NEEDDOC}

  \item{gridDig}{NEEDDOC}

  \item{titleOffsetKoef}{NEEDDOC}

  \item{titleOffsetX}{NEEDDOC}

  \item{minorToMajorKoef}{NEEDDOC}

  \item{sparseKoef}{NEEDDOC}

  \item{graphWidth}{NEEDDOC}

  \item{graphHeight}{NEEDDOC}

  \item{gridCex}{NEEDDOC}

  \item{labelCex}{NEEDDOC}

  \item{titles}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

