\name{easyRect}
\alias{easyRect}
\title{Plot Rectangle}
\usage{
  easyRect(startXY, xLen, yLen, corner = TRUE, ...)
}
\arguments{
  \item{startXY}{NEEDDOC}

  \item{xLen}{NEEDDOC}

  \item{yLen}{NEEDDOC}

  \item{corner}{NEEDDOC}

  \item{\dots}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

