\name{freqReport}
\alias{freqReport}
\title{Frequency Report}
\usage{
  freqReport(type, panel, treat, longPanel = panel,
    typeLabel = label(type), plotprop = FALSE,
    Ntreat = NULL, omitZeros = TRUE, ylim = c(0, 1), h = 5,
    w = 5, digits = 3, size = NULL, longtable = FALSE,
    lines.page = 40, append = FALSE)
}
\arguments{
  \item{type}{NEEDDOC}

  \item{panel}{NEEDDOC}

  \item{treat}{NEEDDOC}

  \item{longPanel}{NEEDDOC}

  \item{typeLabel}{NEEDDOC}

  \item{plotprop}{NEEDDOC}

  \item{Ntreat}{NEEDDOC}

  \item{omitZeros}{NEEDDOC}

  \item{ylim}{NEEDDOC}

  \item{h}{NEEDDOC}

  \item{w}{NEEDDOC}

  \item{digits}{NEEDDOC}

  \item{size}{NEEDDOC}

  \item{longtable}{NEEDDOC}

  \item{lines.page}{NEEDDOC}

  \item{append}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

