\name{makeTreatKey}
\alias{makeTreatKey}
\title{Make Treatment Key}
\usage{
  makeTreatKey(tlevels, paramFile, append = FALSE)
}
\arguments{
  \item{tlevels}{vector. unique treatment levels, expected
  to have length two}

  \item{paramFile}{character. params filename}

  \item{append}{logical.  If \sQuote{TRUE} output will be
  appended to \sQuote{file}, otherwise, contents will be
  overwritten}
}
\description{
  Use treatment levels to generate treatment key in LaTeX.
}
\details{
  Typically will be called with \code{paramFile} missing as
  the filename will be generated by package options.
}
\examples{
makeTreatKey(c('A', 'B'), paramFile='') # prints to standard output
}

