\name{pkgOptions}
\alias{pkgOptions}
\alias{rreport.options}
\title{RReport Package Options}
\description{
  RReport Package Options
}
\section{Options used in rreport}{
  \describe{ \item{\code{rreport.gtype}:}{graphing device
  (ps, pdf, interactive)}
  \item{\code{rreport.appendix.file.name}:}{filename for
  appendix}
  \item{\code{rreport.generated.tex.dir}:}{directory name
  for report tex files}
  \item{\code{rreport.graphics.dir}:}{directory name for
  report graphic files}
  \item{\code{rreport.closed.filename.mask}:}{mask for
  closed report filenames}
  \item{\code{rreport.open.filename.mask}:}{mask for open
  report filenames} }
}
\seealso{
  \code{\link[base]{options}}
}

