\name{plotKmHalfCL}
\alias{plotKmHalfCL}
\title{Plot K-M Half-Width Confidence Levels}
\usage{
  plotKmHalfCL(fit, times, fun = function(x) x, offset = 0,
    lwd = 0.5, lty = 1)
}
\arguments{
  \item{fit}{NEEDDOC}

  \item{times}{NEEDDOC}

  \item{fun}{NEEDDOC}

  \item{offset}{NEEDDOC}

  \item{lwd}{NEEDDOC}

  \item{lty}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

