\name{plotVect}
\alias{plotVect}
\title{Plot Vector}
\usage{
  plotVect(vect, xlim = NULL, ylim = NULL, newGraph = TRUE,
    plotGrid = TRUE, gridVal = NULL, gridCex = 0.5,
    startXYCoor = c(0, 0), vertical = TRUE,
    alignment = c(1, 2), maxLen = 1, minLen = 0.1,
    width = 0.1, breakLen = 0.5,
    labels = as.character(vect), labelMaxLen = 15,
    labelCex = 0.5, col = gray(0.8), lwd = NULL,
    title = "", titleCex = 1, titleXY = startXYCoor)
}
\arguments{
  \item{vect}{NEEDDOC}

  \item{xlim}{NEEDDOC}

  \item{ylim}{NEEDDOC}

  \item{newGraph}{NEEDDOC}

  \item{plotGrid}{NEEDDOC}

  \item{gridVal}{NEEDDOC}

  \item{gridCex}{NEEDDOC}

  \item{startXYCoor}{NEEDDOC}

  \item{vertical}{NEEDDOC}

  \item{alignment}{NEEDDOC}

  \item{maxLen}{NEEDDOC}

  \item{minLen}{NEEDDOC}

  \item{width}{NEEDDOC}

  \item{breakLen}{NEEDDOC}

  \item{labels}{NEEDDOC}

  \item{labelMaxLen}{NEEDDOC}

  \item{labelCex}{NEEDDOC}

  \item{col}{NEEDDOC}

  \item{lwd}{NEEDDOC}

  \item{title}{NEEDDOC}

  \item{titleCex}{NEEDDOC}

  \item{titleXY}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

