\name{publishPdf}
\alias{publishPdf}
\title{Publish PDF}
\usage{
  publishPdf(reports, minutes = NULL, title, server, path,
    extension = "pdf", upload = TRUE, email = FALSE,
    uid = NULL, passwd = NULL, to = NULL, cc = NULL,
    bcc = NULL, sig = NULL, hardcopies = TRUE,
    verbose = TRUE, mailer = c("mail", "kmail"),
    extra = NULL)
}
\arguments{
  \item{reports}{NEEDDOC}

  \item{minutes}{NEEDDOC}

  \item{title}{NEEDDOC}

  \item{server}{NEEDDOC}

  \item{path}{NEEDDOC}

  \item{extension}{NEEDDOC}

  \item{upload}{NEEDDOC}

  \item{email}{NEEDDOC}

  \item{uid}{NEEDDOC}

  \item{passwd}{NEEDDOC}

  \item{to}{NEEDDOC}

  \item{cc}{NEEDDOC}

  \item{bcc}{NEEDDOC}

  \item{sig}{NEEDDOC}

  \item{hardcopies}{NEEDDOC}

  \item{verbose}{NEEDDOC}

  \item{mailer}{NEEDDOC}

  \item{extra}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

