\name{putFig}
\alias{putFig}
\title{Put Figure}
\usage{
  putFig(panel, name, caption = NULL, longcaption = NULL,
    append = TRUE)
}
\arguments{
  \item{panel}{NEEDDOC}

  \item{name}{NEEDDOC}

  \item{caption}{NEEDDOC}

  \item{longcaption}{NEEDDOC}

  \item{append}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

