\name{rangeCheck}
\alias{rangeCheck}
\title{Range Check}
\usage{
  rangeCheck(data, colheader = "Variable", panel,
    append = FALSE)
}
\arguments{
  \item{data}{NEEDDOC}

  \item{colheader}{NEEDDOC}

  \item{panel}{NEEDDOC}

  \item{append}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

