\name{getReferenceObject}
\alias{generateRef}
\alias{getLabelsByKey}
\alias{getReference}
\alias{getReferenceObject}
\alias{getReferenceString}
\alias{getRefsByKey}
\alias{print.latexReference}
\alias{putReferenceObject}
\alias{updateMarkers}
\title{Reference Objects}
\usage{
  getReferenceObject()

  putReferenceObject(refD)

  print.latexReference(refD)

  updateMarkers(newMarker, keyword = "", label = "")

  generateRef()

  getReference(keyword = "", label = "")

  getRefsByKey(keyword = NULL)

  getLabelsByKey(keyword = NULL)

  getReferenceString(keyword)
}
\arguments{
  \item{refD}{}

  \item{newMarker}{}

  \item{keyword}{}

  \item{label}{}
}
\description{
  summary
}
\details{
  details
}
\examples{
1
1
1
1
1
1
1
1
1
}

