\docType{package}
\name{rreport}
\alias{package-rreport}
\alias{rreport}
\alias{rreport-package}
\title{This package creates reports.}
\description{
  This package creates reports.
}
\author{
  Frank E Harrell Jr \email{f.harrell@vanderbilt.edu}

  Maintainer: Charles Dupont
  \email{charles.dupont@vanderbilt.edu}
}

