\name{rreportInit}
\alias{appendixName}
\alias{AppendixName}
\alias{AppendixPath}
\alias{closedGraphDir}
\alias{closedNameMask}
\alias{closedTexDir}
\alias{FilenameMask}
\alias{GraphDirName}
\alias{openGraphDir}
\alias{openNameMask}
\alias{openTexDir}
\alias{rreportInit}
\alias{TexDirName}
\title{Initialize R Report}
\usage{
  rreportInit(dir.open.tex, dir.closed.tex, dir.open.graph,
    dir.closed.graph, open.mask, closed.mask,
    appendix.name, empty.dir = FALSE)

  openNameMask(filename, mask)

  closedNameMask(filename, mask)

  FilenameMask(filename, open.report)

  openGraphDir(name)

  closedGraphDir(name)

  GraphDirName(open.report)

  openTexDir(name)

  closedTexDir(name)

  TexDirName(open.report)

  appendixName(name)

  AppendixName(open.report)

  AppendixPath(open.report)
}
\arguments{
  \item{dir.open.tex}{NEEDDOC}

  \item{dir.closed.tex}{NEEDDOC}

  \item{dir.open.graph}{NEEDDOC}

  \item{dir.closed.graph}{NEEDDOC}

  \item{open.mask}{NEEDDOC}

  \item{closed.mask}{NEEDDOC}

  \item{appendix.name}{NEEDDOC}

  \item{empty.dir}{NEEDDOC}

  \item{filename}{NEEDDOC}

  \item{mask}{NEEDDOC}

  \item{open.report}{NEEDDOC}

  \item{name}{NEEDDOC}
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

