\name{completenessReport}
\alias{completenessReport}
\title{Completeness Report}
\usage{
  completenessReport(data, vars, panel, Time, times,
    longPanel = panel, frac = 0.95,
    h = 4 * (nv < 15) + 5 * (nv >= 15), cex = 1,
    fullCaption = FALSE, append = TRUE)
}
\arguments{
  \item{data}{NEEDDOC}

  \item{vars}{NEEDDOC}

  \item{panel}{NEEDDOC}

  \item{Time}{NEEDDOC}

  \item{times}{NEEDDOC}

  \item{longPanel}{NEEDDOC}

  \item{frac}{NEEDDOC}

  \item{h}{NEEDDOC}

  \item{cex}{NEEDDOC}

  \item{fullCaption}{NEEDDOC}

  \item{append}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

