\name{complianceReport}
\alias{complianceReport}
\title{Compliance Report}
\usage{
  complianceReport(comply, treat, time, times = NULL)
}
\arguments{
  \item{comply}{NEEDDOC}

  \item{treat}{NEEDDOC}

  \item{time}{NEEDDOC}

  \item{times}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

