\name{dropoutReport}
\alias{dropoutReport}
\title{Dropout Report}
\usage{
  dropoutReport(d.dropout, dropout, treat, time.inc = NULL,
    ylim = c(0, 1), panel = "dropout", what = "study", ...)
}
\arguments{
  \item{d.dropout}{NEEDDOC}

  \item{dropout}{NEEDDOC}

  \item{treat}{NEEDDOC}

  \item{time.inc}{NEEDDOC}

  \item{ylim}{NEEDDOC}

  \item{panel}{NEEDDOC}

  \item{what}{NEEDDOC}

  \item{\dots}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

