\name{labReport}
\alias{labReport}
\title{Lab Report}
\usage{
  labReport(data, vars, panel, treat, id, time, times,
    longPanel = panel, h = 6.5, w = 6.5, diffs = FALSE,
    cdfPlot = length(levels(data[[treat]])) <= 2,
    tables = TRUE,
    boxPlot = is.logical(cdfPlot) && !cdfPlot, digits = 3,
    append = FALSE, clearPlots = FALSE, open = TRUE, ...)
}
\arguments{
  \item{data}{NEEDDOC}

  \item{vars}{NEEDDOC}

  \item{panel}{NEEDDOC}

  \item{treat}{NEEDDOC}

  \item{id}{NEEDDOC}

  \item{time}{NEEDDOC}

  \item{times}{NEEDDOC}

  \item{longPanel}{NEEDDOC}

  \item{h}{NEEDDOC}

  \item{w}{NEEDDOC}

  \item{diffs}{NEEDDOC}

  \item{cdfPlot}{NEEDDOC}

  \item{tables}{NEEDDOC}

  \item{boxPlot}{NEEDDOC}

  \item{digits}{NEEDDOC}

  \item{append}{NEEDDOC}

  \item{clearPlots}{NEEDDOC}

  \item{open}{NEEDDOC}

  \item{\dots}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

