\name{listTable}
\alias{listTable}
\title{List Table}
\usage{
  listTable(fileName, longtable = TRUE, landscape = FALSE,
    caption = "", fontSize = "small", dataframe,
    zebraPattern = "none", by = names(dataframe)[1],
    orderGroups = FALSE, colNames = names(dataframe),
    vars = names(dataframe), fixedColVars = c(),
    fixedColWdths = c(), markVar = "", markVarVal = "",
    toLatexChar = TRUE, appendix = TRUE, subsection = NULL,
    marker = NULL, append = FALSE)
}
\arguments{
  \item{fileName}{NEEDDOC}

  \item{longtable}{NEEDDOC}

  \item{landscape}{NEEDDOC}

  \item{caption}{NEEDDOC}

  \item{fontSize}{NEEDDOC}

  \item{dataframe}{NEEDDOC}

  \item{zebraPattern}{NEEDDOC}

  \item{by}{NEEDDOC}

  \item{orderGroups}{NEEDDOC}

  \item{colNames}{NEEDDOC}

  \item{vars}{NEEDDOC}

  \item{fixedColVars}{NEEDDOC}

  \item{fixedColWdths}{NEEDDOC}

  \item{markVar}{NEEDDOC}

  \item{markVarVal}{NEEDDOC}

  \item{toLatexChar}{NEEDDOC}

  \item{appendix}{NEEDDOC}

  \item{subsection}{NEEDDOC}

  \item{marker}{NEEDDOC}

  \item{append}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

