\name{mockListing}
\alias{mockListing}
\title{Mock Listing}
\usage{
  mockListing(control, caption, fields, orientation = "",
    size = "")
}
\arguments{
  \item{control}{NEEDDOC}

  \item{caption}{NEEDDOC}

  \item{fields}{NEEDDOC}

  \item{orientation}{NEEDDOC}

  \item{size}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

