\name{pkgOptions}
\alias{pkgOptions}
\alias{rreport.options}
\title{RReport Package Options}
\description{
  RReport Package Options
}
\section{Options used in rreport}{
  \describe{ \item{\code{rreport.gtype}:}{graphing device
  (ps, pdf, interactive)}
  \item{\code{rreport.appendix.file.name}:}{filename for
  appendix}
  \item{\code{rreport.closed.generated.tex.dir}:}{directory
  name for closed report tex files}
  \item{\code{rreport.open.generated.tex.dir}:}{directory
  name for open report tex files}
  \item{\code{rreport.closed.graphics.dir}:}{directory name
  for closed report graphic files}
  \item{\code{rreport.open.graphics.dir}:}{directory name
  for open report graphic files}
  \item{\code{rreport.closed.filename.mask}:}{mask for
  closed report filenames}
  \item{\code{rreport.open.filename.mask}:}{mask for open
  report filenames} }
}
\seealso{
  \code{\link[base]{options}}
}

