\name{plotEvents}
\alias{plotEvents}
\title{Plot Events}
\usage{
  plotEvents(major, minor, stratLevels, width = 0.5,
    breakWidth = 0.5, sparseKoef = 1.5,
    minorToMajorKoef = 5, plotGrid = TRUE,
    majorGrid = NULL, minorGrid = NULL, title = "",
    fileName = NULL, titleOffsetKoef = 3, titleOffsetX = 0,
    graphWidth = 8, graphHeight = 11, gridCex = 0.5,
    labelCex = 0.5, labelLen = 10)
}
\arguments{
  \item{major}{NEEDDOC}

  \item{minor}{NEEDDOC}

  \item{stratLevels}{NEEDDOC}

  \item{width}{NEEDDOC}

  \item{breakWidth}{NEEDDOC}

  \item{sparseKoef}{NEEDDOC}

  \item{minorToMajorKoef}{NEEDDOC}

  \item{plotGrid}{NEEDDOC}

  \item{majorGrid}{NEEDDOC}

  \item{minorGrid}{NEEDDOC}

  \item{title}{NEEDDOC}

  \item{fileName}{NEEDDOC}

  \item{titleOffsetKoef}{NEEDDOC}

  \item{titleOffsetX}{NEEDDOC}

  \item{graphWidth}{NEEDDOC}

  \item{graphHeight}{NEEDDOC}

  \item{gridCex}{NEEDDOC}

  \item{labelCex}{NEEDDOC}

  \item{labelLen}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

