\name{putFig}
\alias{putFig}
\title{Put Figure}
\usage{
  putFig(panel, name, caption = NULL, longcaption = NULL,
    append = TRUE)
}
\arguments{
  \item{panel}{character. Panel name.}

  \item{name}{character. Name for figure.}

  \item{caption}{character. Short caption for figure.}

  \item{longcaption}{character. Long caption for figure.}

  \item{append}{logical. If \sQuote{TRUE} output will be
  appended instead of overwritten.}
}
\description{
  Included a generated figure within LaTex document.
}

