\name{repVarclus}
\alias{repVarclus}
\title{Variable Clustering Diagrams}
\usage{
  repVarclus(data, time, times = sort(unique(time)),
    nmin = 10)
}
\arguments{
  \item{data}{NEEDDOC}

  \item{time}{NEEDDOC}

  \item{times}{NEEDDOC}

  \item{nmin}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

