\name{floor.chron}
\alias{ceiling.chron}
\alias{floor.chron}
\title{Round Chronological Objects}
\usage{
  floor.chron(x,
    units = c("minutes", "hours", "days", "months", "years"))

  ceiling.chron(x,
    units = c("minutes", "hours", "days", "months", "years"))
}
\arguments{
  \item{x}{NEEDDOC}

  \item{units}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
1
}

