\name{subjectList}
\alias{subjectList}
\title{Subject List}
\usage{
  subjectList(data, panel, caption = NULL,
    vname = c("labels", "names"), colheads = NULL,
    size = "smaller", longtable = TRUE, landscape = TRUE)
}
\arguments{
  \item{data}{NEEDDOC}

  \item{panel}{NEEDDOC}

  \item{caption}{NEEDDOC}

  \item{vname}{NEEDDOC}

  \item{colheads}{NEEDDOC}

  \item{size}{NEEDDOC}

  \item{longtable}{NEEDDOC}

  \item{landscape}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

