\name{survReport}
\alias{survReport}
\title{Survival Report}
\usage{
  survReport(etime, event, treat, group = NULL,
    treat.group = NULL, data,
    ylabel = "Survival Probability",
    conf = c("bars", "bands", "none"), n = NULL,
    labels = NULL, previousActual = NULL, h = 4,
    append = FALSE, fileName = "trt",
    descrip = "treatment", ndescrip = "", ...)
}
\arguments{
  \item{etime}{NEEDDOC}

  \item{event}{NEEDDOC}

  \item{treat}{NEEDDOC}

  \item{group}{NEEDDOC}

  \item{treat.group}{NEEDDOC}

  \item{data}{NEEDDOC}

  \item{ylabel}{NEEDDOC}

  \item{conf}{NEEDDOC}

  \item{n}{NEEDDOC}

  \item{labels}{NEEDDOC}

  \item{previousActual}{NEEDDOC}

  \item{h}{NEEDDOC}

  \item{append}{NEEDDOC}

  \item{fileName}{NEEDDOC}

  \item{descrip}{NEEDDOC}

  \item{ndescrip}{NEEDDOC}

  \item{\dots}{NEEDDOC}
}
\value{
  return something
}
\description{
  summary
}
\details{
  details
}
\examples{
1
}

